refactor: unifica dois hooks em um só
