+++ apps/frontend/src/app/(pages)/(internal)/(customers)/myAppointments/page.tsx
@@ -2,11 +2,11 @@
 import InputDay from "@/components/appointments/InputDay";
 import ItemAppointment from "@/components/appointments/ItemAppointment";
 import Header from "@/components/shared/Header";
-import useUserSchedule from "@/data/hooks/useUserSchedule";
+import useLoadSchedule from "@/data/hooks/useLoadSchedule";
 import { IconCalendarCancel } from "@tabler/icons-react";
 
 export default function PageMyAppointments(){
-    const {appointments, changedDate, date, deleteAppointment} = useUserSchedule()
+    const {appointments, changedDate, date, deleteAppointment} = useLoadSchedule()
     return(
         <section className="flex flex-col bg-zinc-900">
             <Header title="Minha Agenda" description="Visualize seus agendamentos" />
